requirements.txt
streamlit==1.49.1
python-docx==1.2.0
openpyxl==3.1.5
Pillow==11.0.0
